Update tests, add more granularity to errors.
