+++ Teacher/Teacher/SpeedGrader/StudentPager/Model/SpeedGraderInteractor.swift
@@ -18,17 +18,6 @@
 
 import Combine
 import Core
-import Foundation
-
-enum SpeedGraderInteractorState {
-    case loading
-    case loaded(
-        assignment: Assignment,
-        submissions: [Submission],
-        focusedSubmissionIndex: Int
-    )
-    case error(Error)
-}
 
 protocol SpeedGraderInteractor {
     var state: CurrentValueSubject<SpeedGraderInteractorState, Never> { get }
@@ -40,96 +29,20 @@ protocol SpeedGraderInteractor {
     func refreshSubmission(forUserId: String)
 }
 
-let SpeedGraderAllUsersUserID = "speedgrader"
-
-class SpeedGraderInteractorLive: SpeedGraderInteractor {
-    let state = CurrentValueSubject<SpeedGraderInteractorState, Never>(.loading)
-
-    public let assignmentID: String
-    public let userID: String
-    public let context: Context
-
-    private let env: AppEnvironment
-    private let filter: [GetSubmissions.Filter]
-    private var subscriptions = Set<AnyCancellable>()
-
-    init(
-        env: AppEnvironment,
-        context: Context,
-        assignmentID: String,
-        userID: String,
-        filter: [GetSubmissions.Filter]
-    ) {
-        self.env = env
-        self.context = context
-        self.assignmentID = assignmentID
-        self.userID = userID
-        self.filter = filter
-    }
-
-    func loadInitialData() {
-        loadAssignment()
-            .flatMap { [weak self] assignment in
-                guard let self else {
-                    return Fail<(Assignment, [Submission]), Error>(error: NSError.internalError())
-                        .eraseToAnyPublisher()
-                }
-                return Publishers.CombineLatest(
-                    loadEnrollments(),
-                    loadSubmissions(anonymizeStudents: assignment.anonymizeStudents)
-                )
-                .map { (assignment, $0.1) }
-                .eraseToAnyPublisher()
-            }
-            .sink { [weak self] completion in
-                if case .failure(let error) = completion {
-                    self?.state.send(.error(error))
-                }
-            } receiveValue: { [weak self] (assignment: Assignment, submissions: [Submission]) in
-                guard let self else { return }
-                let focusedSubmissionIndex = submissions.firstIndex { self.userID == SpeedGraderAllUsersUserID || $0.userID == self.userID }
-                let data = SpeedGraderInteractorState.loaded(
-                    assignment: assignment,
-                    submissions: submissions,
-                    focusedSubmissionIndex: focusedSubmissionIndex ?? 0
-                )
-                self.state.send(data)
-            }
-            .store(in: &subscriptions)
-    }
-
-    func refreshSubmission(forUserId: String) {
-        let submissionUseCase = GetSubmission(context: context, assignmentID: assignmentID, userID: userID)
-        ReactiveStore(useCase: submissionUseCase, environment: env)
-            .getEntities(ignoreCache: true)
-            .sink()
-            .store(in: &subscriptions)
-    }
-
-    private func loadAssignment() -> AnyPublisher<Assignment, Error> {
-        let assignmentUseCase = GetAssignment(
-            courseID: context.id,
-            assignmentID: assignmentID,
-            include: [.overrides]
-        )
-        return ReactiveStore(useCase: assignmentUseCase, environment: env)
-            .getEntities()
-            .tryMap { try $0.first.unwrapOrThrow() }
-            .eraseToAnyPublisher()
-    }
-
-    private func loadSubmissions(anonymizeStudents: Bool) -> AnyPublisher<([Submission]), Error> {
-        let submissionsUseCase = GetSubmissions(context: context, assignmentID: assignmentID, filter: filter)
-        submissionsUseCase.shuffled = anonymizeStudents
-        return ReactiveStore(useCase: submissionsUseCase, environment: env)
-            .getEntities(loadAllPages: true)
-            .eraseToAnyPublisher()
-    }
-
-    private func loadEnrollments() -> AnyPublisher<[Enrollment], Error> {
-        let enrollmentsUseCase = GetEnrollments(context: context)
-        return ReactiveStore(useCase: enrollmentsUseCase, environment: env)
-            .getEntities(loadAllPages: true)
-            .eraseToAnyPublisher()
-    }
+enum SpeedGraderInteractorState {
+    case loading
+    case data(
+        assignment: Assignment,
+        submissions: [Submission],
+        focusedSubmissionIndex: Int
+    )
+    case error(SpeedGraderInteractorError)
 }
+
+enum SpeedGraderInteractorError: Error {
+    case userIdNotFound
+    case submissionNotFound
+    case unexpectedError(Error)
+}
+
+let SpeedGraderAllUsersUserID = "speedgrader"
